# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TaskBackend API
  version: 1.1.0
  description: Production-hardened TaskBackend API with JWT auth, RBAC, soft deletion, and observability endpoints.
servers:
  - url: http://localhost:3000
paths:
...
  /api/products:
    get:
      summary: List products
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10, maximum: 100 }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: [createdAt_desc, createdAt_asc, price_asc, price_desc]
      responses:
        '200':
          description: Product list
    post:
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductInput:
      type: object
      required: [title, price]
      properties:
        title: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }